% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomad-package.R
\docType{package}
\name{lomad-package}
\alias{lomad-package}
\title{lomad: Local Moving Average Decoupling}
\description{
Detects and characterizes periods of decoupling between two time series using correlations between local moving averages.
}
\references{
Ruiz, T. D., Seifert, A. J., Hamilton, E., Mispagel, C. M., Hunt, O. P.,
Garcia, J., and Bockmon, E. E. (2026). Inference for local trend similarity
in nonstationary time series via rolling correlation, with application to
assessing stability in an estuarine system. Manuscript in preparation.

Hall, P. and Van Keilegom, I. (2003). Using difference-based methods for
inference in nonparametric regression with time series errors. \emph{Journal
of the Royal Statistical Society Series B}, 65(2), 443--456.

Benjamini, Y. and Yekutieli, D. (2001). The control of the false discovery
rate in multiple testing under dependency. \emph{The Annals of Statistics},
29(4), 1165--1188.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ruizt/lomad-package}
  \item Report bugs at \url{https://github.com/ruizt/lomad-package/issues}
}

}
\author{
\strong{Maintainer}: Trevor Ruiz \email{truiz01@calpoly.edu}

Authors:
\itemize{
  \item Trevor Ruiz \email{truiz01@calpoly.edu}
  \item Alea Seifert
  \item Catherine Mispagel
  \item Emma Hamilton
  \item Andres Rocha
}

}
\keyword{internal}
